+++ ims/include/huaweicloud/ims/v2/model/ListImagesRequest.h
@@ -40,7 +40,7 @@ public:
     /// ListImagesRequest members
 
     /// <summary>
-    /// 镜像类型，目前支持以下类型： 公共镜像：gold 私有镜像：private 共享镜像：shared
+    /// 镜像类型，目前支持以下类型： 公共镜像：gold 私有镜像：private 共享镜像：shared 市场镜像：market
     /// </summary>
 
     std::string getImagetype() const;